Generic functions ifex_ast_introspect -> ast_utils
There were more generic functions in ifex_ast_introspect.  The functions
to build/investigate/manipulate AST trees are general for all similar
ASTs so they are now also moved to the common utils file.

There may be a few duplicates for what we already moved from
ifex_ast_construction file, but they can be cleaned up later.
