+++ output_filters/schema/ifex_to_json_schema.py
@@ -15,17 +15,15 @@ Generate JSON Schema equivalent to the python-internal model definition.
 
 from dataclasses import fields
 from datetime import datetime, date
-
+from models.common.ast_utils import actual_type, field_actual_type, field_inner_type, inner_type, is_forwardref, is_list, type_name, field_is_list, field_is_optional, field_referenced_type
 from models.common.ast_utils import is_simple_type
-
 from models.ifex.ifex_ast import AST
-from models.ifex.ifex_ast_introspect import actual_type, field_actual_type, field_inner_type, inner_type, is_forwardref, is_list, type_name, field_is_list, field_is_optional, field_referenced_type
 from typing import Any
 import json
 
 # =======================================================================
 # Helpers
-# (Most general ones are in ifex_ast_introspect module instead)
+# (Most general ones are in ast_utils module instead)
 # =======================================================================
 
 # This function is local because "string" and "integer" are neither python
